extract workspace docker build panel
